Move FileUserMetadata to storages plugin
